Use Netgen Layouts permissions to check access in controllers

+++ bundles/BlockManagerAdminBundle/Controller/Admin/LayoutResolver/DeleteCondition.php
@@ -7,7 +7,7 @@ namespace Netgen\Bundle\BlockManagerAdminBundle\Controller\Admin\LayoutResolver;
 use Netgen\BlockManager\API\Service\LayoutResolverService;
 use Netgen\BlockManager\API\Values\LayoutResolver\Condition;
 use Netgen\BlockManager\View\ViewInterface;
-use Netgen\Bundle\BlockManagerAdminBundle\Controller\Admin\Controller;
+use Netgen\Bundle\BlockManagerBundle\Controller\Controller;
 
 final class DeleteCondition extends Controller
 {
@@ -26,6 +26,8 @@ final class DeleteCondition extends Controller
      */
     public function __invoke(Condition $condition): ViewInterface
     {
+        $this->denyAccessUnlessGranted('nglayouts:mapping:edit');
+
         $this->layoutResolverService->deleteCondition($condition);
 
         return $this->buildView(
