+++ bundles/BlockManagerAdminBundle/Controller/Admin/LayoutResolver/DeleteTarget.php
@@ -7,7 +7,7 @@ namespace Netgen\Bundle\BlockManagerAdminBundle\Controller\Admin\LayoutResolver;
 use Netgen\BlockManager\API\Service\LayoutResolverService;
 use Netgen\BlockManager\API\Values\LayoutResolver\Target;
 use Netgen\BlockManager\View\ViewInterface;
-use Netgen\Bundle\BlockManagerAdminBundle\Controller\Admin\Controller;
+use Netgen\Bundle\BlockManagerBundle\Controller\Controller;
 
 final class DeleteTarget extends Controller
 {
@@ -26,6 +26,8 @@ final class DeleteTarget extends Controller
      */
     public function __invoke(Target $target): ViewInterface
     {
+        $this->denyAccessUnlessGranted('nglayouts:mapping:edit');
+
         $this->layoutResolverService->deleteTarget($target);
 
         return $this->buildView(
