+++ bundles/BlockManagerBundle/Controller/API/V1/Collection/DeleteItems.php
@@ -6,7 +6,7 @@ namespace Netgen\Bundle\BlockManagerBundle\Controller\API\V1\Collection;
 
 use Netgen\BlockManager\API\Service\CollectionService;
 use Netgen\BlockManager\API\Values\Collection\Collection;
-use Netgen\Bundle\BlockManagerBundle\Controller\API\Controller;
+use Netgen\Bundle\BlockManagerBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Response;
 
 final class DeleteItems extends Controller
@@ -26,6 +26,8 @@ final class DeleteItems extends Controller
      */
     public function __invoke(Collection $collection): Response
     {
+        $this->denyAccessUnlessGranted('nglayouts:collection:items');
+
         $this->collectionService->deleteItems($collection);
 
         return new Response(null, Response::HTTP_NO_CONTENT);
